Move Expunge out of the account object
