+++ legacy/account/src/main/java/app/k9mail/legacy/account/Expunge.kt
@@ -1,0 +1,16 @@
+package app.k9mail.legacy.account
+
+import com.fsck.k9.backend.api.SyncConfig.ExpungePolicy
+
+enum class Expunge {
+    EXPUNGE_IMMEDIATELY,
+    EXPUNGE_MANUALLY,
+    EXPUNGE_ON_POLL,
+    ;
+
+    fun toBackendExpungePolicy(): ExpungePolicy = when (this) {
+        EXPUNGE_IMMEDIATELY -> ExpungePolicy.IMMEDIATELY
+        EXPUNGE_MANUALLY -> ExpungePolicy.MANUALLY
+        EXPUNGE_ON_POLL -> ExpungePolicy.ON_POLL
+    }
+}
