+++ legacy/common/src/main/java/com/fsck/k9/backends/ImapBackendFactory.kt
@@ -2,8 +2,8 @@ package com.fsck.k9.backends
 
 import android.content.Context
 import app.k9mail.legacy.account.Account
-import app.k9mail.legacy.account.Account.Expunge
 import app.k9mail.legacy.account.AccountManager
+import app.k9mail.legacy.account.Expunge
 import com.fsck.k9.backend.BackendFactory
 import com.fsck.k9.backend.api.Backend
 import com.fsck.k9.backend.imap.ImapBackend
